tast-tests: Migrate tests to use cryptohome.{Downloads,MyFiles}Path
The usage of the /home/chronos/user/Downloads bind mount is being
deprecated. This moves a chunk of tests away from using the
filesapp.DownloadPath and filesapp.MyFilesPath and to directly
retrieving the mount via cryptohome. The tests affected in this patch
are:

  - ui.DMICRecord*
  - network.CertSettingsPage*
  - scanapp.OpenScanInFilesApp*
  - scanapp.MultiPageScan*
  - typec.HotplugUSBHeadsetAudioPlayback*

This also updates the scanapp methods that run app settings tests to
use cryptohome method MyFilesPath as well.

BUG=b:233132576
TEST=tast run all tests listed above
TEST=./fast_build.sh -T

Change-Id: Iec8c56fe4997a6ddc216cf3d79fbd1d8499f5e21
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3733937
Commit-Queue: Ben Reich <benreich@chromium.org>
Tested-by: Ben Reich <benreich@chromium.org>
Reviewed-by: Jun Sheng <shengjun@chromium.org>