tast-tests: Migrate helpers to use cryptohome.{Downloads,MyFiles}Path
The usage of the /home/chronos/user/Downloads bind mount is being
deprecated. This moves a chunk of tests away from using the
filesapp.DownloadPath and filesapp.MyFilesPath and to directly
retrieving the mount via cryptohome. The tests affected in this patch
are:

  - policy.DisableScreenshotsHotkey*
  - policy.DisableScreenshotsCaptureMode*
  - dlp.DataLeakPreventionRulesListScreenshot*
  - notifications.ScreenCaptureNotification*
  - notifications.NotificationExpandCollapse*

This also updates some helper methods to plumb through (or retrieve) the
path instead of relying on the hardcoded mount point.

BUG=b:233132576
TEST=tast run all tests listed above
TEST=./fast_build.sh -T

Change-Id: Ifd621c377dc73f97e51c79d271b5200734e58ebc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3733938
Reviewed-by: Jun Sheng <shengjun@chromium.org>
Reviewed-by: Oleh Lamzin <lamzin@google.com>
Tested-by: Ben Reich <benreich@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>