Improve stability of DragToShowOverview
Adjust the formula for calculating how long to swipe from the bottom of
a screen in tablet mode to enter overview mode. The initial method of
calculating the swipe duration led to swipes that were too short, and
sometimes not registered. For example, on a machine with a display height of 864 pixels, the swipe duration was 288ms, which is less than the hard-coded 300ms used in DragToShowHomescreen. Thus, some calls to DragToShowOverview would either not register the swipe at all, or accidentally land the DUT into the homescreen page. By increasing the duration of the swipe, while keeping the duration a function of the height of the display, this change increases stability for average-sized devices, while not harming performance on larger ones.

BUG=b:236752651
TEST=Run tast run <DUT> ui.VideoCUJ.tablet ui.VideoCUJ.lacros_tablet,
and compare error rates of this function before and after this change.
Initially, when trying to use this function for 10 minutes consecutively, it would hit a failure within a couple minutes. Now, it
can succeed throughout the entire duration of the test.

Change-Id: Ifa6c0945c666703a4006285c5f6653f6474e6de0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3736565
Tested-by: Navin Ramsaroop <ramsaroop@google.com>
Commit-Queue: Navin Ramsaroop <ramsaroop@google.com>
Reviewed-by: Avery Musbach <amusbach@chromium.org>
Reviewed-by: Yichen Zhou <yichenz@chromium.org>
Reviewed-by: Jun Sheng <shengjun@chromium.org>