security.SELinuxProcesses*: Clarify failure message.
We've heard from developers and sheriffs that the failure message for
this test is not clear. Make it better.

-Update the name of the domain we check for from "minijailed" to
"unconfined_proc". "minijailed" gave the impression of things being
"OK" ("jailed" must be OK, right?) which made the test failures
confusing. In reality "minijailed" was the domain assigned to things
launched from the minijail0 executable that don't have their own
domain. We want everything to have its own domain so a "minijailed"
domain is a bad thing. Having the "bad" domain be called "minijailed"
was confusing. The new name ("unconfined_proc") at least alludes to
the process being unconfined, which is a bad thing.
The domain is renamed in the dependent CL.

-Update the String() method for the Process struct so that the cmdline
comes before the SELinux domain.

-Update the failure message to explicitly say that the process needs
its own SELinux domain rather than saying "want anything except", where
"anything" is not very clear.

-Update the final error message to clarify that the failure is a
security bug, and update the link to a better section in the SELinux
documentation.

BUG=b:237445387
TEST=Run a process in an unconfined domain, run the test, check the
TEST=error message.

Cq-Depend: chromium:3733433
Change-Id: Iee8812b80a4f5ca9132cee7acd7511a9a6a13689