Add basic Instant Tethering test
Add a Tast test for basic Instant Tethering functionality. The test
assumes the connected phone has a mobile data connection. The test will
disable WiFi and ethernet on the Chromebook, which triggers the Instant
Tethering connection popup, which we'll use to connect the Chromebook to
the phone's mobile data.

BUG=b:237696234
TEST=tast run <dut> instanttether.Basic
TEST=test_that localhost:2222 tast.cross-device.local --autotest_dir ~/chromiumos/src/third_party/autotest/files --args="phone_station=localhost phone_station_ssh_port=2222 android_serial=99151FFBA00410"

Change-Id: If0e198270a0ebe9df7be7bc6a2d2ab7ac81fb661
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3739250
Reviewed-by: Ryan Hansberry <hansberry@chromium.org>
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>
Commit-Queue: Kyle Shimabukuro <kyleshima@chromium.org>