tast-tests: add Onscreen to settings nodes
The node could be cached in the ui tree without being visible on screen.
(i.e., http://b/237595911)

That said, checking whether a node exists in the UI tree is not the same
as checking the node exists on screen.

Add Onscreen to the crostini UI settings UI nodes.

BUG=b:237595911
TEST=tast run <DUT> crostini.*

Change-Id: If2b4c9a3d04475969762a10e97c2db1fc0fc35f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3736244
Reviewed-by: David Munro <davidmunro@google.com>
Tested-by: Alvin Jia <alvinjia@google.com>
Commit-Queue: Alvin Jia <alvinjia@google.com>