lacros.Migrate: Prevent too frequent page reload.
I have observed locally with atlas, which is a slower device, that tast
often fails to find an UI element due to the page being reloaded before
the element gets the chance to be rendered.

This change either gets rid of the reload or slower the pace of reload
so that Chrome have enough time to render the elements test is looking
for.

BUG=chromium:1340820
TEST=tast run $DUT lacros.Migrate.*

Change-Id: I4f63e180f3b2eaf6a5d55cbcc5e9eec6804b364a