faft: Give a better error message
If there is no ui.signinProfileTestExtensionManifestKey, this test
fails. This makes for mysterious failures for the users, so instead
check explicitly for this missing var and point the users to
documentation on how to fix it. Also, since this moves the failure to
later in the test, the test can be skipped if the HardwareDeps don't
match even if the var is missing.

BUG=b:237609801,b:237660482
TEST=Ran test in chroot without tast-tests-private

Cq-Depend: chromium:3737349
Change-Id: I50141991ed1ed551deaa42e20962962d6a70b66c