tast: dnsproxy: Fix doh-automatic
DoH automatic upgrade logic is based on a hard-coded mapping of
well-known nameservers to DoH provider URLs. The test lab uses
a custom vlan with it's own nameservers, so the lookup fails and
there is no DoH provider to use.

This patch manually set the DoH mapping to the test device's
nameserver.

Alongside that, this patch also allows retries for the automatic
test. This is necessary because the failure timeout for the proxy
might be longer than the client's.

BUG=b:235429092
TEST=lease a DUT through crosfleet;
     tast run <DUT> network.DNSProxy.doh_automatic

Change-Id: I3443741355c8d8fe656ade779dfe70ddb74a4572
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3740068
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Commit-Queue: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>