network: Parameterize shill_captive_portal_http.go tast test
This builds upon the shill_captive_portal_http.go tast test and
paramaterizes it to also include a test for the 'portal-suspected' state
when the httpserver redirect has no location in the header. The
HTTPResponseHandler function can now be passed from the test to
customize the handler.

BUG=b:237811539
TEST=tast run ${DUT} network.ShillCaptivePortalHTTP*

Change-Id: I3eda623e76af816cb12b2e29a647a126828b370f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3739230
Tested-by: Michael Rygiel <michaelrygiel@google.com>
Commit-Queue: Michael Rygiel <michaelrygiel@google.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>