tast-tests: Restore ARCVM conf if bootedFixture.SetUp() fails.
While debugging b:237334804, I noticed that if the fixture for
arc.VideoDecodeAccelVD.* fails, the arcvm_dev.conf is not restored. This
may cause the test for the next fixture to fail if the configuration
that's left in arcvm_dev.conf is problematic.

While bootedFixture does restore the arcvm_dev.conf in TearDown(), this
method does not run if SetUp() fails [1]:

  // If one or more errors are reported in SetUp by s.Error or s.Fatal,
  // all remaining tests depending on this fixture are marked failed
  // without actually running. TearDown is not called in this case. If
  // s.Fatal is called, SetUp immediately aborts.

This CL fixes this by making sure RestoreArcvmDevConf() is called when
SetUp() fails.

[1] https://source.chromium.org/chromiumos/chromiumos/codesearch/+/main:src/platform/tast/src/chromiumos/tast/internal/testing/fixt.go;l=209-211;drc=b351bca65c7e25adf4181397108f662570207182

BUG=b:237334804
TEST=tast run ${drallion} arc.VideoDecodeAccelVD.h264_vm arc.UnicornParentPermission.vm

Change-Id: I693bd81efccf59b7027b35624189fd5defaa7700
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3740169
Reviewed-by: Renuga Nanjappan <rnanjappan@chromium.org>
Tested-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Renuga Nanjappan <rnanjappan@chromium.org>