tast-tests: Adopt new API to detect IME readiness on Zhuyin
A new test API was introduced recently to detect IME readiness for testing: chrome.autotestPrivate.isInputMethodReadyForTesting.

All IME related tests should be gradually migrated to use this API instead of hard sleep.

Why start with Zhuyin: This API only works on physical keyboard at the moment. Zhuyin is currently only used for PK testing and also not in CQ (informational test).

Note: There is no way to know the test uses PK or VK in the lib function. It is possible to explicitly call different function in the test script, which can complicate the solution.

BUG=b:237955874
TEST=tast run <DUT> inputs.PhysicalKeyboardZhuyinTyping

Change-Id: Ic16a30810400d102de1a2e07caa5f2187fbcf964