u2fd: Wait 1 second before clicking Retry button
In the recent 100% test failures of this test, it seems like the "Try
again" button exists, but the DoDefault doesn't have any effect. We
indeed wait until it exists before pressing it.

When local testing, I can't 100% reproduce it but can reproduce it with
20~30% chance. I guess the button is not clickable right after
appearing, so added a 1-second sleep before clicking it, and the test
never fails locally anymore. I think we can observe whether this fix
completely resolves the problem.

BUG=b:237840070
TEST=tast run $DUT u2fd.WebauthnPINLockout for many times

Change-Id: I1187194ee36038392defbec7d28eaff621706cd3