u2fd: Click Retry button after it's focused
In the recent 100% test failures of this test, it seems like the "Try
again" button exists, but the DoDefault doesn't have any effect. We
did wait until it exists before pressing it.

When local testing, I can't 100% reproduce it but can reproduce it with
20~30% chance. I guess the button is not clickable right after
appearing, so adding EnsureFocused logic to ensure that it's focused
before we click it.

BUG=b:237840070
TEST=tast run $DUT u2fd.WebauthnPINLockout for many times

Change-Id: I1187194ee36038392defbec7d28eaff621706cd3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3742809
Tested-by: Howard Yang <hcyang@google.com>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Yi Chou <yich@google.com>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Howard Yang <hcyang@google.com>