tast-tests: add test for VK overscroll
Tast test is added to check that VK overscroll feature works correctly
in Ash and Lacros. The tests include the following user interaction
scenarios:

* Clicking on an input field at the bottom of a page: The test checks
  that overscroll is performed to prevent the input field to be
  overlaid by VK given that the both are at the bottom of the page.
* Typing with VK in a text field outside the viewport: This one also
  checks overscroll is performed to make the input field visible.

Notes: All the tests are initially set as "informational" to allow
first verifying their correctness on different devices and
environments.

Important Note: Lacros implementation of overscroll is ongoing. As a
result, relevant tests are failing at the moment.

BUG=b:237354374
TEST=tast run <DUT> inputs.VirtualKeyboardOverscroll.tablet
TEST=tast run <DUT> inputs.VirtualKeyboardOverscroll.clamshell

Change-Id: Id34fe528bde17906c0efd14272e75332f533167d