tast: Add app compatibility test for CodeMirror.
CodeMirror is a popular Javascript library for code editors [1] used by ChromeOS
Text app and various other apps.

Add a simple test for editing CodeMirror text with the virtual keyboard.

For now, we'll use a CDN to grab the CodeMirror asset files. Once we
figure out how to correctly host the CodeMirror assets locally, we can
switch to a local version for more stability.

[1] https://codemirror.net/

BUG=b:237495943
TEST=tast run <dut> inputs.AppCompatibility*

Change-Id: Ia40f86754b5fc25076632db0eff90b266ce61fe0