network: Configure prefixlen in EthernetStaticIP test
When configuring StaticIPParameters, prefixlen and address should be
set at the same time. This affects the logic not only in the current
implementation in shill (StaticIPParameters::ContainsAddress() checks
prefixlen as well), but also after the ongoing Network refactor (address
without prefix is not valid).

This change also reveals another race issue in the test. See the
comments in waitForAddressCacheInShill() for details.

BUG=b:232177767
TEST=tast run $DUT network.EthernetStaticIP
TEST=../platform/tast/fast_build.sh -T

Change-Id: I1ef1429dbca8dab47f8b958424f68f180bcabc62
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3746236
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Jun Sheng <shengjun@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>