tast-tests: Fix multi matches of finder with ClassName TimeView
There are now two TimeView nodes visible due to the UI changes of the
status area. The name of the nodes are identical too, just one shows the
date and one shows the time.

Therefore, only the first match from the finder would be used to verify
the time format.

BUG=chromium:1342185
TEST=tast run <ip> lockscreen.DisplayProperTimeFormat*

Change-Id: If1028cd6dd40c252c87fa31d8818e0fdc65d7415