tast-test: Fix a labelling issue with cryptohome tests
Make default label to "gaia" in cryptohome tests. This label is used to
test Vault creation, which is expected to happen with new vaults.

Additionally, modifies the arc test to unmount before remounting. This
makes the test more robust as the user is expected to unmount before the
update is applied.

BUG=b:237815647
FIXED=b:237815647
TEST=Cq Passes

Change-Id: I8a9218b7629a9f591d541afaab51b0ef0b042912