fingerprint: Reverse the isFPUpdaterEnabled logic
The isFPUpdaterEnabled function checks for the existence of the
'.disable_fp_updater' file which actually suggests that the FP updater
is disabled.
On Guybrush and Zork devices, the mixup caused the file
'.disable_fp_updater' to be kept and thus prevented the execution of the 'FpUpdater' test

Reversed the logic to return the correct Enabled value.

BUG=b:233892105
TEST=tast -verbose run < IP > firmware.FpAddEntropy

Change-Id: I98330c53fb007009e833eafb1f571d2d9b018883
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3750181
Tested-by: Ofer Fried <oferfried@google.com>
Commit-Queue: Ofer Fried <oferfried@google.com>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Patrick Meiring <meiring@google.com>
Reviewed-by: Patryk Duda <patrykd@google.com>