fingerprint: Reverse the isFPUpdaterEnabled logic
The isFPUpdaterEnabled function checks for the existance of the
'.disable_fp_updater' file which actually suggests that the FP updater
is disabled.
On Guybrush and Zork devices, the mixup caused the file
'.disable_fp_updater' to be kept and thus prevented the execution of the 'FpUpdater' test

Reversed the logic to return isFPUpdaterDisabled instead.

BUG=b:233892105
TEST=tast -verbose run < IP > firmware.FpAddEntropy

Change-Id: I98330c53fb007009e833eafb1f571d2d9b018883