tast-tests: Change appsplatform.WebAPK to use Files SWA
The WebAPK system in Chrome is migrating to use a new feature which only
works when the Files SWA is enabled. The Files SWA is enabled by default
in Chrome but disabled in tast tests. This explicitly enables it for
appsplatform.WebAPK so that the test continues to work.

Also fixes an incorrect error message while I'm here

BUG=chromium:1342129
TEST=tast run $DUT appsplatform.WebAPK

Change-Id: I7e526e8bd6f35d5b07cd80c81405d6a7a4f7ba16
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3750810
Reviewed-by: Alvin Jia <alvinjia@google.com>
Auto-Submit: Tim Sergeant <tsergeant@chromium.org>
Tested-by: Tim Sergeant <tsergeant@chromium.org>
Commit-Queue: Tim Sergeant <tsergeant@chromium.org>