Revert "crash: Mark MountFailure test informational"
This reverts commit eaccb2630f2077b578f9883811a5f4cb02b08a98.

Reason for revert: Was failing because apps.SystemWebAppsReinstall was inducing a crash on the reven-vmtest board. That test has since been marked informational so this test should pass again. We have b/238173644 open to improve the crash. tasts' handling of the case where chrome has crashed and session manager is down.

Original change's description:
> crash: Mark MountFailure test informational
>
> BUG=b:237696882
> TEST=CQ
>
> Change-Id: Ic0c2f0df5d0ce573aa2aa8d102aee5ac0b875651
> Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3738523
> Auto-Submit: Joseph Sussman <josephsussman@google.com>
> Reviewed-by: Seewai Fu <seewaifu@google.com>
> Tested-by: Joseph Sussman <josephsussman@google.com>

Bug: b:237696882
Change-Id: I60242ec9adba6fc39f78a1a841e9b90b46699451