tast-tests: fix issue in removing crostini
The correct steps of removing crostini are:
1, Enter Linux development env subpage.
2, Click Remove, and the RemoveConfirmDialog pops up.
3, Click Delete.
4, Wait for the RemoveLinuxAlert gone.

In step 3, the currently logic is to click "Delete" and wait until
RemoveLinuxAlert exists, which doesn't make sense, as the expected
results is that the alert dialog is gone in step 4.

This causes the test to fail due to errors in b/238376419
if the uninstallation takes less than 5 seconds.

BUG=b:238376419
TEST=tast run <DUT> crostini.*

Change-Id: Id46c1a8f75ef520854876eb469f3fb23a6365dd2