tast-tests: Use ecryptfs in CrossVersionLogin
Tell cryptohome to create users with ecryptfs when preparing snapshots
for hwsec.CrossVersionLogin. This has to be done in two places: one is
to pass "--ecryptfs" to cryptohome CLI, and another is to pass
"--cryptohome-use-old-encryption-for-testing" to Chrome. This way, both
password-based user and challenge-response-based user are covered.

Note that "--cryptohome-use-old-encryption-for-testing" was added in
Chrome M-103, so when creating snapshots from older versions ecryptfs
won't be used for password users by default. This will need to be
addressed separately.

The reason for enabling ecryptfs is that it allows the test to take
snapshots of encrypted user vaults: the
hwsec.PrepareCrossVersionLoginData will pack snapshots created on the
developer machine, and hwsec.CrossVersionLogin will use them when
running the test (e.g., on bots). Ecryptfs supports backups of encrypted
files, unlike, e.g., ext4 fscrypt (the current default used by
cryptohome). Follow-up CLs will implement taking advantage of these user
vault snapshots by implementing remount/relogin test scenarios.

BUG=b:237120336
TEST=run prepare_cross_version_login_data.sh on M-90 with tpm2-simulator, tpm2, crrev.com/c/3610151 backported, check the created config.json has "HasUserVault: true" for both users, untar the created .tar.gz, check it has files like "/home/.shadow/<...>/vault/user/ECRYPTFS_FNEK_ENCRYPTED.<...>"

Change-Id: I58349a7788ce0a4e780ea29406ce58db8abc22cc