tast-tests: fix issue in removing crostini
The correct steps of removing crostini are:
1, Enter Linux development env subpage.
2, Click Remove, and the RemoveConfirmDialog pops up.
3, Click Delete.
4, Wait for the RemoveLinuxAlert gone.

In step 3, the currently logic is to click "Delete" and wait until
RemoveLinuxAlert exists, which may appear and disappear quickly before
the check "WaitUntilExists".

This causes the test to fail due to errors in b/238376419
if the uninstallation is fast and the alert dialog only occurs for a
very short time.

Remove the check for the alert dialog and change step 4 to
directly wait for the "turn on" button.
Use "DoDefault" to replace LeftClick, as it is more stable.

BUG=b:238376419
TEST=tast run <DUT> crostini.*

Change-Id: Id46c1a8f75ef520854876eb469f3fb23a6365dd2