tast-tests: fix issue in removing crostini
The correct steps of removing crostini are:
1, Enter Linux development env subpage.
2, Click Remove, and the RemoveConfirmDialog pops up.
3, Click Delete.
4, Wait for the RemoveLinuxAlert gone.

In step 3, the currently logic is to click "Delete" and wait until
RemoveLinuxAlert exists, which may appear and disappear quickly before
the check "WaitUntilExists".

This causes the test to fail due to errors in b/238376419
if the uninstallation is fast and the alert dialog only occurs for a
very short time. This is when the test is in a state where
RemoveConfirmDialog.Delete and RemoveLinuxAlert both do not exist.

Instead, we can just wait until the RemoveConfirmDialog is gone.

BUG=b:238376419
TEST=tast run <DUT> crostini.*

Change-Id: Id46c1a8f75ef520854876eb469f3fb23a6365dd2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3749875
Reviewed-by: Sophia Lin <sophialin@google.com>
Reviewed-by: David Munro <davidmunro@google.com>
Commit-Queue: Alvin Jia <alvinjia@google.com>
Tested-by: Alvin Jia <alvinjia@google.com>