tast-tests: stabilize crostini.AppGeditFileSharing
This test is flaky on some devices due to the screen size/resolution.

The test tries to detect the full sentence "This is a text string in
a text file in the Downloads folder.", which may not be fully detected
by the OCR model.

Change it to a partial query "text string" to verify the content is
rendered.

BUG=b:236312054
TEST=tast run <DUT> crostini.AppGeditFilesharing*

Change-Id: I0ca2e4935f5d0746936ec567be9eb4b880ede5d1