wificell: Remove technical debt from TestFixture
2nd phase of TestFixture refactor - spinning off new objects
(WiFiDUT and WiFiRouter) to simplify fixture operations.
The decoupled objects don't need TestFixture proxying anymore
to be used.
Moreover, the objects contain some parts of their interface (WiFiHost),
that is common, so they can be used in more abstract manner
(e.g. to send ping between arbitrary hosts).

In this phase we keep TestFixture's API intact, that will be changed in
the third phase.

BUG=b:234845693
TEST=tast run $DUT '("group:wificell" && wificell_func && \
  !wificell_unstable && "name:wifi.*")'

Change-Id: I7ee13bf429e6d5f9515cd59cc1ac1f06ede230fc