tast-tests: Add LaunchSWAToPath and use in DrivefsDssOffline
The DriveFS tests rely on a pool of 10 accounts which regularly ends up
having multiple tests (or DUTs) using the same login and thus there ends
up being congestion when trying to add files to the root directory.
Splitting these out into unique folders per test and run is good, but
all methods require navigating to the folder via clicking. When other
tests are running the UI is not stable and new folders can appear
asynchronously.

To avoid this race condition, a new autotestPrivate API was introduced
that launches Files app directly to the absolute path supplied. This
allows a unique folder to be created unique to the current test and test
run, then open Files app directly to this folder. This avoids congestion
on navigation AND in the root directory. Fow now this just updates the
filemanager.DrivefsDssOffline test to this new paradigm, but follow up
CLs will migrate all Drive tests.

BUG=chromium:1299712
TEST=tast run DUT filemanager.DrivefsDssOffline

Change-Id: I48fbd4e894ecbc9ccdff8e1cd189a01132e82c91
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3752778
Tested-by: Ben Reich <benreich@chromium.org>
Reviewed-by: Austin Tankiang <austinct@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>