crash: Make test resilient to prior chrome crashes.
Adds a check to make sure that the UI job has started
in order to prevent tests from failing due to `daemon-store`
being unavailable.

This is causes by chrome crashing in a prior test. Waiting
for the UI job to restart ensures that chrome has finished
restarting and that `daemon-store` will be available.

BUG=b:238173644
TEST=tast run <DUT> apps.SystemWebAppsReinstall crash.MountFailure

Change-Id: Ia48e977906599607bea49dfb75805174bf4a20dc