launcher: Generalize launcher test setup
Most of the launcher tests have boilerplate code during setup to set
tablet mode state, open launcher, and wait for the set of apps to
stabilize. This creates a helper method to set up launcher test case
to launcher.go that performs common set up steps.

The setup method returns a cleanup function intended to be called by
the test that ensures that the tablet mode is reset, and that the
launcher is closed after the test. Logic to close the launcher in
deferred function is intended to minimize the chance of launcher tests
leave system UI in a state with the launcher open, which did happen with
few existing tests).

While going over tests, make them more consistent in shortening context
used in tests, and make sure that "dump on error" deferred calls happen
after deferred calls to reset tablet mode / launcher state.

BUG=chromium:1329277
TEST=tast run launcher.*

Change-Id: Iaf020bb37b3e9f0297aff9a8457d97f2d1a21a32