launcher: Generalize launcher test setup
Most of the launcher tests have boilerplate code during setup to set
tablet mode state, open launcher, and wait for the set of apps to
stabilize. This creates a helper method to set up launcher test case
to launcher.go that performs common set up steps.

The setup is adapted to better handle the case when the launcher is open
when the test starts (in most cases tests were just clicking the home
button to open the launcher, which has an opposite effect if the
launcher is already open at that point). Additionally, the setup method
returns a cleanup function intended to be called by the test that
ensures that the launcher is closed after the test (minimizing the
chance that tests leave system UI in a state with the launcher open).

BUG=chromium:1329277
TEST=tast run launcher.*

Change-Id: Iaf020bb37b3e9f0297aff9a8457d97f2d1a21a32