tast-tests: Add lacros variant for appsplatform.WebAPK
This test interacts with a web app, which means that we need to install
and launch the app in the primary browser.

Installing web apps currently uses an autotestPrivate API. These APIs
only work in ash-chrome, so instead we use uiauto to click through the
install buttons in the browser.

BUG=chromium:1310143
TEST=tast run $DUT appsplatform.WebAPK appsplatform.WebAPK.lacros \
  apps.Sharesheet

Change-Id: I1d8e652570219797f8df2bfd8e7d18e9025709aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3878734
Reviewed-by: Ben Reich <benreich@chromium.org>
Commit-Queue: Tim Sergeant <tsergeant@chromium.org>
Tested-by: Tim Sergeant <tsergeant@chromium.org>