tast-tests: Capture details of TPM init errors
Attach additional error details to the TPM initialization errors, so
that it's possible to understand which exactly command failed.

These errors are considered non-critical for Tast, still it's often the
case that the tests will break if the TPM is not owned. Investigating
such issues is much easier if Tast tells you at which step the
initialization aborted.

BUG=b:237120336
TEST=start an M91 VM that doesn't support "tpm_manager_client get_supported_features", run any Tast test, check that the "Failed ensuring that TPM is initialized" message contains details on which command has failed

Change-Id: I746a662d73072daf134f7fbbccadf8d0be9085c3