arc.Print: Move models to stable test
We originally moved a bunch of models to arc.Print.unstable so we could
run arc.Print in the CQ on stable boards.  Since then, a bunch more
boards are consistently passing, so let's move them out of the unstable
test.  This should leave us with just a few unstable boards, so it will
be much easier to investigate the remaining failures and potentially
remove the unstable variant entirely.

BUG=b:238638672
TEST=CQ

Change-Id: I2feae31aeea3cff22c04074ed3deb24dda73c216
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3754611
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Piotr Pawliczek <pawliczek@chromium.org>
Reviewed-by: Mohamed Omar <mohamedaomar@google.com>