ui: Try for higher outbound video resolution in ui.MeetCUJ.*
A bot with the spotlight layout will be added to the meeting, and it
will pin the test user to potentially elicit higher video quality from
the DUT, depending on what the device and the network can adequately
support. If the meeting_code test variable is used to run the test with
a preexisting meeting, then the test will add the bot with the spotlight
layout to that meeting.

BUG=b:231750257
TEST=Verify that on a powerful device with good network conditions, this
CL increases the following metrics in results-chart.json for MeetCUJ.*:
WebRTCInternals.Video.Outbound.frameHeight
WebRTCInternals.Video.Outbound.frameWidth

Change-Id: I8e495da9ea2f8ebf9fbf43796261a43033505b5e