tast: add timeout between input focus and keypress
As outlined in b/235417796, there appears to be a race condition
occurring when the its.ValidateInputOnField method is used to validate
test input.

This seems to be more apparent with lacros browsers, as there is more
time taken for a focus event to be received in ash (and consequently
just that little bit longer to fully activate an input method).

Adding a 50ms delay between detecting an input as focused, and entering
the first keypress seems to solve the issue.

BUG=b:235417796
TEST=tast -verbose run -build=true 127.0.0.1:9222
inputs.PhysicalKeyboardTransliterationTyping.el_lacros

Change-Id: Ibe524ed969d924eeb6cf9ec2a0575673ea6e7256
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3759804
Reviewed-by: Alvin Jia <alvinjia@google.com>
Reviewed-by: Darren Shen <shend@chromium.org>
Reviewed-by: Jun Sheng <shengjun@chromium.org>
Commit-Queue: Curtis McMullan <curtismcmullan@chromium.org>
Tested-by: Darren Shen <shend@chromium.org>