tast-tests: move restartIfError to fixture
The Apps tests restart Crostini if the tests fail. This is to make sure
the open apps are closed to leave a clean env for the following tests,
and may cause unexpected failures if not.
I.e. in https://storage.cloud.google.com/chromeos-test-logs/test-runner/prod/2022-07-12/cf171a30-6c06-4a0e-b2fc-f9ccf138bd6f/autoserv_test/tast/results/tests/crostini.AppVscode.tablet_unstable/uidetection_screenshot.png
the Gedit is not closed when testing VS code.

It is redundant to do so in every individual tests, and can be omitted.

It makes more sense to move this logic to PostTest in the fixture.

BUG=b:231258823
TEST=tast run <DUT> crostini.App*

Change-Id: If284003c13da278260acfbca391a44a16a2bde3c