tast-tests: stabilize GeditUnshareFolder
Add a verification to wait until the Gedit app is fully loaded before
proceeding to further steps.

The test is flaky in tablet mode due to a race condition between loading
the app and loading the settings page.

In the actions of opening a file in Gedit, the test only waits for the
window node to appear in a11y tree, then it immediately tries to open
the settings page.

It could enter the situation that the Gedit window is fully loaded after
loading the settings page then the test fails to find nodes in settings.

Clamshell mode could have the similar problem, but since the app is not
in the maximized state, the app and the settings page can co-exist,
which is impossible in the tablet mode.

BUG=b:238841934
TEST=tast run <DUT> crostini.AppGeditUnshareFolder.tablet_unstable

Change-Id: I0746e2811ad1036853ad188b76b08a72072bf13b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3759817
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Commit-Queue: Alvin Jia <alvinjia@google.com>
Tested-by: Alvin Jia <alvinjia@google.com>