camera: Add ArcCameraApp test to cover basic tests for ARC camera app
Originally, we install GCA to exercise the basic camera functionalities
of an ARC camera app. However, since we no longer maintain GCA, this CL
leverages the existing ArcCameraFpsTest app which lives in Tast repo to
exercise the same functionalities that camera.GCA* tests do.

Once this test is stable, we can remove camera.GCA* tests and totally
remove the maintainance of GCA on Google3 repo.

BUG=b:238836004
TEST=tast run {DUT} camera.ArcCameraApp

Change-Id: I5116699248989b1887acb33feae6709a3800a9f9