android: Fix release lock several times issue in ArcCameraFpsTest
In ArcCameraFpsTest app, if we trigger SET_CAMERA_ID first then trigger
START_RECORDING, sometimes it will hit an exception which is caused by a
race between closeCamera() and startPreview().

The reason is that CameraCaptureSession.StateCallback.onConfigured()
might be triggered more than once for a capture session, but every time
the onConfigured() or onConfigureFailed() is triggered, we increase the
semaphore by one. This CL fixes this issue by only increase semaphore if
it is currently zero.

BUG=b:238836004
TEST=tast run {DUT} camera.ArcCameraApp (with descendent CL)

Change-Id: I62565581983a9b27c04f9dab90d81fdbe6f5b59f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3759818
Tested-by: Wei Lee <wtlee@chromium.org>
Commit-Queue: Wei Lee <wtlee@chromium.org>
Reviewed-by: Ben Reich <benreich@chromium.org>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>