firmware: Add CorruptBothFWBodyAB() test case
This test using servo v4+ to corrupt the firmware sections containing
body. The CorruptBothFWBodyAB() test requires a plugged-in flash
drive with a test image, this test corrupts both firmware signature A
and B. On next reboot, the firmware verification fails and enters
recovery mode. This test then checks the success of the recovery boot.

This test replaces firmware_CorruptBothFwBodyAB.py from FAFT.

BUG=b:194907767
TEST=tast run -var servo=${SERVO_IP} ${DUT_IP} firmware.CorruptBothFWBodyAB.normal_mode
TEST=tast unit test passed.

Change-Id: Idcefbfde8e14e0687503ba935f8a549286ee4c09