arc.Print: Launch print preview from tast instead of Android
The current test app behavior is to launch print preview as soon as the
activity is created.  This causes print preview to be rapidly opened and
canceled as the app is resized, which may be the cause of some rare
flakes.

Instead of launching print preview immediately, put a button on the main
app activity and make tast click it.  This avoids unnecessary print
preview launches. This should also make the WaitForPrintPreview check
more accurate because it doesn't race against the app startup.

BUG=b:159325335
TEST=40 iterations of `tast run arc.Print.*` on grunt and hatch

Change-Id: Idd3adba38fa49bfe8840de318c0169a2c832fd83
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3760792
Reviewed-by: Hiroki Sato <hirokisato@chromium.org>
Reviewed-by: Bryan Cain <bryancain@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Alvin Jia <alvinjia@google.com>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>