tast-test: Cleanup ARC install without using settings
Currently, we clean up the test by uninstalling an app using the
Settings app. When a test fails, the Settings app is opened and it could
affect following tests by giving preference to the Settings app instead
of the newest install.

ARC API has methods for uninstalling an app package, so defer the
cleanup function to use those instead of the settings app.

BUG=chromium:1344547
TEST=tast run <DU>T launcher.RecentApps.android*

Change-Id: Id10360d98b781ad3d46f45eef9fdd8eaadba4892