network: Add HTTPS functionality to http virtualnet package
This builds upon the virtualnet package and allows the creation of an
https server. The captive portal tests use the https server to get a
valid 204 response which will cause the service state to be changed to
online.

In order to create a valid https server, we must install the certs to
the appropriate file location. This is done using a bind mount to the
test certs in a tmp directory.

BUG=b:237816536
TEST=tast run ${DUT} network.ShillCaptivePortalHTTP*

Change-Id: Ie122936e2aba1275e50d475fee11c93f793e0c1f