tast-tests: stabilize crostini.RestartApp
This test is flaky in chrome postsubmit due to reasons in b/238955541.
The remaining problem is about Gedit not properly being closed.

WaitUntilExists(appWindow) doesn't necessarily mean that the crostini
app is launched, i.e., due to racing issues, the app could be loaded
after appear in a11y tree.

Add a verification of the app being launched by checking if the Open
button is rendered on screen.

BUG=b:238955541
TEST=tast run <DUT> crostini.RestartApp

Change-Id: I498083758c5fa061c8954b5593b05d86c76bb615