tast-tests: stabilize crostini.RestartApp
This test is flaky in chrome postsubmit due to reasons in b/238955541.
The remaining problem is about Gedit not properly being closed.

WaitUntilExists(appWindow) doesn't necessarily mean that the crostini
app is launched, i.e., due to racing issues, the app could be loaded
after appearing in a11y tree.

Add a verification to check if the Open button is rendered on screen as
indicator of the app being loaded.

BUG=b:238955541
TEST=tast run <DUT> crostini.RestartApp

Change-Id: I498083758c5fa061c8954b5593b05d86c76bb615