firmware: Add TryFWB() test case
This test uses servo v4+ to switch into the FW A section if it was in
FW B. And then set a flag to try boot from the FW B section. On the
next reboot, the test verifies that the boot into FW B was successful.
In the end, test perform one more reboot and check boot into correct
FW variant (B if vboot2 else A).

This test replaces firmware_TryFwB.py from FAFT.

BUG=b:194910847
TEST=tast run -var servo=${SERVO_IP} ${DUT_IP} firmware.TryFWB.normal_mode

Change-Id: Ifbdef32e3bfaa96ba21cfc58ff936ce7518a80af