scanapp: Remove extra printer.Stop calls
We originally shut down the virtual printer without waiting for the
defer stack to unwind so that tests wouldn't block on background Chrome
processing.  The scanapp.* tests have been changed to wait for Chrome to
finish processing the printer before starting the scan app, so this
extra shutdown is no longer needed.  Remove the extra printer.Stop calls
and add some logs to make sure we can tell where the test stopped in
case this reintroduces a problem.

BUG=b:210134772
TEST=tast run '("name:scanapp.*" && "group:mainline")'

Change-Id: I1616c2c7f1c804c7c43589323e6004eaf2df9c90