usbprinter: Fix udevadm race
We run udevadm as a subprocess and read events from it in a goroutine.
However, we previously didn't wait for the child process to be started
before proceeding with adding or removing the virtual USB printer.  If
the child process didn't start quickly enough, this could cause us to
miss a udev event, which causes an otherwise-successful test to fail.

To avoid this race, add another channel to waitEvent and make it send
the child process ID once udevadm is ready.  Then the caller can read
from this channel to synchronize before manipulating the printer.

BUG=b:230787720
TEST=100 iterations of tast run '("dep:virtual_usb_printer")'

Change-Id: I0487a1c660041c20d1f6e5456e98a8247f00b9e3